renaming and restructuring testing modules for consistency
