+++ core/src/ledger/governance/storage/proposal.rs
@@ -313,14 +313,14 @@ impl Display for StorageProposal {
 }
 
 #[cfg(any(test, feature = "testing"))]
-/// Tests and strategies for governance proposals
-pub mod tests {
+/// Testing helpers and and strategies for governance proposals
+pub mod testing {
     use proptest::prelude::Strategy;
     use proptest::{collection, option, prop_compose};
 
     use super::*;
     use crate::types::address::testing::arb_non_internal_address;
-    use crate::types::hash::tests::arb_hash;
+    use crate::types::hash::testing::arb_hash;
     use crate::types::token::testing::arb_amount;
 
     /// Generate an arbitrary add or removal of what's generated by the supplied
