+++ core/src/types/hash.rs
@@ -148,15 +148,28 @@ impl From<Hash> for crate::tendermint::Hash {
 
 #[cfg(any(test, feature = "testing"))]
 /// Tests and strategies for hashes
+pub mod testing {
+    use proptest::prop_compose;
+
+    use super::*;
+
+    prop_compose! {
+        /// Generate an arbitrary hash
+        pub fn arb_hash()(bytes: [u8; 32]) -> Hash {
+            Hash(bytes)
+        }
+    }
+}
+
+#[cfg(test)]
+/// Tests and strategies for hashes
 pub mod tests {
     use proptest::prelude::*;
-    #[cfg(test)]
     use proptest::string::{string_regex, RegexGeneratorStrategy};
 
     use super::*;
 
     /// Returns a proptest strategy that yields hex encoded hashes.
-    #[cfg(test)]
     fn hex_encoded_hash_strat() -> RegexGeneratorStrategy<String> {
         string_regex(r"[a-fA-F0-9]{64}").unwrap()
     }
@@ -165,13 +178,6 @@ pub mod tests {
         #[test]
         fn test_hash_string(hex_hash in hex_encoded_hash_strat()) {
             let _: Hash = hex_hash.try_into().unwrap();
-        }
-    }
-
-    prop_compose! {
-        /// Generate an arbitrary hash
-        pub fn arb_hash()(bytes: [u8; 32]) -> Hash {
-            Hash(bytes)
         }
     }
 }
